chore: move app list constants to common constants
